Reorganize Backend Code into a Server Folder for Merging Front-end Code
The back-end code previously was sitting at the root of the codebase because it was assumed that front- and back-end would use separate repos. As now both front- and back-end code are placed together in one repo, the back-end code is now placed into a 'server' directory sub to the root to allow front-end code to sit at the root.