Complete User Authenticatoin UI to Merge 'main' to Test API
The revisions included in this commit are based on commit# 6483a03 Base Skeleton and are for the purpose of implementing the authentication UI (user registration/login). The test cannot be conducted without the back-end API in place, however. Yet, the server-sprint-1 prototype back end, though fully funcitonal for the purpose of this stage of front-end development, has yet to be approved and merged. This makes testing any front-end work in connection with back-end API not possible within the 'main' branch. Because the 'server-sprint-1' branch is more than sufficient to provide for testing the authentication UI, this commit is made so the revisions here can be merged into 'server-sprint-1' to test. The revisions here will then be proposed for merging into origin/main once test passes.