Revise 'getAllTodos' controller to 'getTodos' for route '${BASE_URL}/api/v1/todos/?list=${list_id}'
The app's back end currently has APIs that GET all the todo items indiscriminately at sub-route '/todos'. This is not desirable. This revision change the name and behavior of the 'getAllTodos' controller to 'getTodos' for the route '/todos/?list=${list_id}' so only the todo items belonging to a particular list with ${list_id} will be fetched at a time. GET by query string in the route also provide for further revision to include pagination in the GET method at the route.